write match resutls to csv
